added functions for integ test
