+++ query-service/src/integrationTest/resources/attribute-expression-test-queries/query2.json
@@ -1,0 +1,85 @@
+{
+  "filter": {
+    "childFilter": [{
+      "lhs": {
+        "attributeExpression": {
+          "attributeId": "Span.start_time_millis"
+        }
+      },
+      "operator": "GT",
+      "rhs": {
+        "literal": {
+          "value": {
+            "valueType": "LONG",
+            "long": "1570658506605"
+          }
+        }
+      }
+    }, {
+      "lhs": {
+        "attributeExpression": {
+          "attributeId": "Span.start_time_millis"
+        }
+      },
+      "operator": "LT",
+      "rhs": {
+        "literal": {
+          "value": {
+            "valueType": "LONG",
+            "long": "1570744906673"
+          }
+        }
+      }
+    }, {
+      "lhs": {
+        "attributeExpression": {
+          "attributeId": "Span.tags"
+        }
+      },
+      "operator": "CONTAINS_KEY",
+      "rhs": {
+        "literal": {
+          "value": {
+            "string": "span.kind"
+          }
+        }
+      }
+    }, {
+      "lhs": {
+        "attributeExpression": {
+          "attributeId": "Span.tags",
+          "subpath": "span.kind"
+        }
+      },
+      "operator": "NEQ",
+      "rhs": {
+        "literal": {
+          "value": {
+          }
+        }
+      }
+    }]
+  },
+  "selection": [{
+    "function": {
+      "functionName": "AVG",
+      "arguments": [{
+        "attributeExpression": {
+          "attributeId": "Span.duration_millis"
+        }
+      }],
+      "alias": "avg_duration"
+    }
+  }, {
+    "attributeExpression": {
+      "attributeId": "Span.tags",
+      "subpath": "span.kind"
+    }
+  }],
+  "groupBy": [{
+    "attributeExpression": {
+      "attributeId": "Span.tags",
+      "subpath": "span.kind"
+    }
+  }]
+}
